+++ src/Transifex.php
@@ -45,6 +45,7 @@ class Transifex
 	 *
 	 * @var    Formats
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $formats;
 
@@ -53,6 +54,7 @@ class Transifex
 	 *
 	 * @var    LanguageInfo
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $languageinfo;
 
@@ -61,6 +63,7 @@ class Transifex
 	 *
 	 * @var    Languages
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $languages;
 
@@ -69,6 +72,7 @@ class Transifex
 	 *
 	 * @var    Projects
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $projects;
 
@@ -77,6 +81,7 @@ class Transifex
 	 *
 	 * @var    Resources
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $resources;
 
@@ -85,6 +90,7 @@ class Transifex
 	 *
 	 * @var    Statistics
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $statistics;
 
@@ -93,6 +99,7 @@ class Transifex
 	 *
 	 * @var    Translations
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $translations;
 
@@ -101,6 +108,7 @@ class Transifex
 	 *
 	 * @var    Translationstrings
 	 * @since  1.0
+	 * @deprecated  2.0
 	 */
 	protected $translationstrings;
 
@@ -142,6 +150,7 @@ class Transifex
 	 *
 	 * @return  TransifexObject  Transifex API object.
 	 *
+	 * @deprecated  2.0  Use get() method instead
 	 * @since   1.0
 	 * @throws  \InvalidArgumentException
 	 */